Add initial test system
Add a new test space api that behaves slightly differently to the
regular space to allow testing of some aspects of vcycle scheduling
without creating and uploading files etc. It uses a different test
machine class which simulates a vm rather than scanning for one.

At the moment a single cycle and machine creation works but features to
simulate machines have not been added. References to time in shared will
have to be changed in order for it to work in non real time.