Improve preemptive algorithm
Machinetypes now have a holdoff variable which tells preemptible
instances to not be created. This avoids using last abort time and can
be set back to False as soon as high priority machines have aborted,
avoiding large drops in utilisation.

It's currently set to run in cycles, this will be changed later.